feat: [BBND-1498] move Dividend request files from equity to dividend folder
