+++ packages/ats/sdk/__tests__/fixtures/equity/EquityFixture.ts
@@ -13,10 +13,10 @@ import { HederaId } from "@domain/context/shared/HederaId";
 import { GetScheduledBalanceAdjustmentQuery } from "@query/equity/balanceAdjustments/getScheduledBalanceAdjustment/GetScheduledBalanceAdjustmentQuery";
 import { faker } from "@faker-js/faker/.";
 import GetEquityDetailsRequest from "@port/in/request/equity/GetEquityDetailsRequest";
-import SetDividendRequest from "@port/in/request/equity/SetDividendRequest";
-import GetDividendForRequest from "@port/in/request/equity/GetDividendForRequest";
-import GetDividendRequest from "@port/in/request/equity/GetDividendRequest";
-import GetAllDividendsRequest from "@port/in/request/equity/GetAllDividendsRequest";
+import SetDividendRequest from "@port/in/request/dividend/SetDividendRequest";
+import GetDividendForRequest from "@port/in/request/dividend/GetDividendForRequest";
+import GetDividendRequest from "@port/in/request/dividend/GetDividendRequest";
+import GetAllDividendsRequest from "@port/in/request/dividend/GetAllDividendsRequest";
 import SetVotingRightsRequest from "@port/in/request/equity/SetVotingRightsRequest";
 import GetVotingRightsForRequest from "@port/in/request/equity/GetVotingRightsForRequest";
 import GetVotingRightsRequest from "@port/in/request/equity/GetVotingRightsRequest";
@@ -49,7 +49,7 @@ import { CancelScheduledBalanceAdjustmentCommand } from "@command/equity/balance
 import { CreateEquityCommand } from "@command/equity/create/CreateEquityCommand";
 import { SetDividendCommand } from "@command/equity/dividends/set/SetDividendCommand";
 import { CancelDividendCommand } from "@command/equity/dividends/cancel/CancelDividendCommand";
-import CancelDividendRequest from "@port/in/request/equity/CancelDividendRequest";
+import CancelDividendRequest from "@port/in/request/dividend/CancelDividendRequest";
 import { SetVotingRightsCommand } from "@command/equity/votingRights/set/SetVotingRightsCommand";
 import { CancelVotingCommand } from "@command/equity/votingRights/cancel/CancelVotingCommand";
 import CancelVotingRequest from "@port/in/request/equity/CancelVotingRequest";
@@ -59,8 +59,8 @@ import { GetDividendHoldersQuery } from "@query/equity/dividends/getDividendHold
 import { GetTotalDividendHoldersQuery } from "@query/equity/dividends/getTotalDividendHolders/GetTotalDividendHoldersQuery";
 import { GetVotingHoldersQuery } from "@query/equity/votingRights/getVotingHolders/GetVotingHoldersQuery";
 import { GetTotalVotingHoldersQuery } from "@query/equity/votingRights/getTotalVotingHolders/GetTotalVotingHoldersQuery";
-import GetDividendHoldersRequest from "@port/in/request/equity/GetDividendHoldersRequest";
-import GetTotalDividendHoldersRequest from "@port/in/request/equity/GetTotalDividendHoldersRequest";
+import GetDividendHoldersRequest from "@port/in/request/dividend/GetDividendHoldersRequest";
+import GetTotalDividendHoldersRequest from "@port/in/request/dividend/GetTotalDividendHoldersRequest";
 import GetVotingHoldersRequest from "@port/in/request/equity/GetVotingHoldersRequest";
 import GetTotalVotingHoldersRequest from "@port/in/request/equity/GetTotalVotingHoldersRequest";
 import { CreateTrexSuiteEquityCommand } from "@command/equity/createTrexSuite/CreateTrexSuiteEquityCommand";
