+++ packages/ats/sdk/src/port/in/request/index.ts
@@ -18,9 +18,9 @@ import PauseRequest from "./security/operations/pause/PauseRequest";
 import ControlListRequest from "./security/operations/controlList/ControlListRequest";
 import GetControlListCountRequest from "./security/operations/controlList/GetControlListCountRequest";
 import GetControlListMembersRequest from "./security/operations/controlList/GetControlListMembersRequest";
-import GetDividendForRequest from "./equity/GetDividendForRequest";
-import GetDividendRequest from "./equity/GetDividendRequest";
-import GetAllDividendsRequest from "./equity/GetAllDividendsRequest";
+import GetDividendForRequest from "./dividend/GetDividendForRequest";
+import GetDividendRequest from "./dividend/GetDividendRequest";
+import GetAllDividendsRequest from "./dividend/GetAllDividendsRequest";
 import GetVotingRightsForRequest from "./equity/GetVotingRightsForRequest";
 import GetVotingRightsRequest from "./equity/GetVotingRightsRequest";
 import GetAllVotingRightsRequest from "./equity/GetAllVotingRightsRequest";
@@ -34,8 +34,8 @@ import GetRolesForRequest from "./security/roles/GetRolesForRequest";
 import GetRoleMemberCountRequest from "./security/roles/GetRoleMemberCountRequest";
 import GetRoleMembersRequest from "./security/roles/GetRoleMembersRequest";
 import GetSecurityDetailsRequest from "./security/GetSecurityDetailsRequest";
-import SetDividendRequest from "./equity/SetDividendRequest";
-import CancelDividendRequest from "./equity/CancelDividendRequest";
+import SetDividendRequest from "./dividend/SetDividendRequest";
+import CancelDividendRequest from "./dividend/CancelDividendRequest";
 import SetCouponRequest from "./bond/SetCouponRequest";
 import SetVotingRightsRequest from "./equity/SetVotingRightsRequest";
 import GetBondDetailsRequest from "./bond/GetBondDetailsRequest";
@@ -186,8 +186,8 @@ import GetTotalTokenHoldersAtSnapshotRequest from "./security/operations/snapsho
 import BalancesOfAtSnapshotRequest from "./snapshots/BalancesOfAtSnapshotRequest";
 import GetCouponHoldersRequest from "./bond/GetCouponHoldersRequest";
 import GetTotalCouponHoldersRequest from "./bond/GetTotalCouponHoldersRequest";
-import GetDividendHoldersRequest from "./equity/GetDividendHoldersRequest";
-import GetTotalDividendHoldersRequest from "./equity/GetTotalDividendHoldersRequest";
+import GetDividendHoldersRequest from "./dividend/GetDividendHoldersRequest";
+import GetTotalDividendHoldersRequest from "./dividend/GetTotalDividendHoldersRequest";
 import GetTotalVotingHoldersRequest from "./equity/GetTotalVotingHoldersRequest";
 import GetVotingHoldersRequest from "./equity/GetVotingHoldersRequest";
 import GetSecurityHoldersRequest from "./security/GetSecurityHoldersRequest";
